Move the classes and key registration as well.
