+++ src/main/java/com/sk89q/craftbook/sponge/mechanics/blockbags/BlockBagManager.java
@@ -16,10 +16,6 @@
  */
 package com.sk89q.craftbook.sponge.mechanics.blockbags;
 
-import static org.spongepowered.api.data.DataQuery.of;
-import static org.spongepowered.api.data.key.KeyFactory.makeSingleKey;
-
-import com.google.common.reflect.TypeToken;
 import com.google.inject.Inject;
 import com.me4502.modularframework.module.Module;
 import com.me4502.modularframework.module.guice.ModuleConfiguration;
@@ -28,10 +24,7 @@ import com.sk89q.craftbook.core.util.PermissionNode;
 import com.sk89q.craftbook.core.util.documentation.DocumentationProvider;
 import com.sk89q.craftbook.sponge.mechanics.types.SpongeMechanic;
 import com.sk89q.craftbook.sponge.util.SpongePermissionNode;
-import com.sk89q.craftbook.sponge.util.type.TypeTokens;
 import ninja.leaping.configurate.ConfigurationNode;
-import org.spongepowered.api.data.key.Key;
-import org.spongepowered.api.data.value.mutable.Value;
 import org.spongepowered.api.service.permission.PermissionDescription;
 
 import java.util.HashSet;
@@ -40,13 +33,6 @@ import java.util.concurrent.ThreadLocalRandom;
 
 @Module(id = "blockbag", name = "BlockBag", onEnable="onInitialize", onDisable="onDisable")
 public class BlockBagManager extends SpongeMechanic implements DocumentationProvider {
-
-    public static Key<Value<EmbeddedBlockBag>> EMBEDDED_BLOCK_BAG = makeSingleKey(new TypeToken<EmbeddedBlockBag>() {},
-            new TypeToken<Value<EmbeddedBlockBag>>() {}, of("EmbeddedBlockBag"), "craftbook:embeddedblockbag",
-            "EmbeddedBlockBag");
-
-    public static Key<Value<Long>> BLOCK_BAG = makeSingleKey(new TypeTokens.LongTypeToken(),
-            new TypeTokens.LongValueTypeToken(), of("BlockBag"), "craftbook:blockbag", "BlockBag");
 
     public SpongePermissionNode adminPermissions = new SpongePermissionNode("craftbook.blockbag.admin",
             "Allows usage of admin block bags.", PermissionDescription.ROLE_ADMIN);
