+++ src/main/java/com/sk89q/craftbook/sponge/util/data/builder/BlockBagDataManipulatorBuilder.java
@@ -14,9 +14,11 @@
  * You should have received a copy of the GNU General Public License along with this program. If not,
  * see <http://www.gnu.org/licenses/>.
  */
-package com.sk89q.craftbook.sponge.mechanics.blockbags.data;
+package com.sk89q.craftbook.sponge.util.data.builder;
 
-import com.sk89q.craftbook.sponge.mechanics.blockbags.BlockBagManager;
+import com.sk89q.craftbook.sponge.util.data.immutable.ImmutableBlockBagData;
+import com.sk89q.craftbook.sponge.util.data.CraftBookKeys;
+import com.sk89q.craftbook.sponge.util.data.mutable.BlockBagData;
 import org.spongepowered.api.data.DataHolder;
 import org.spongepowered.api.data.DataView;
 import org.spongepowered.api.data.manipulator.DataManipulatorBuilder;
@@ -45,8 +47,8 @@ public class BlockBagDataManipulatorBuilder extends AbstractDataBuilder<BlockBag
 
     @Override
     protected Optional<BlockBagData> buildContent(DataView container) throws InvalidDataException {
-        if (container.contains(BlockBagManager.BLOCK_BAG.getQuery())) {
-            return Optional.of(new BlockBagData(container.getLong(BlockBagManager.BLOCK_BAG.getQuery()).get()));
+        if (container.contains(CraftBookKeys.BLOCK_BAG.getQuery())) {
+            return Optional.of(new BlockBagData(container.getLong(CraftBookKeys.BLOCK_BAG.getQuery()).get()));
         }
 
         return Optional.empty();
