+++ src/main/java/com/sk89q/craftbook/sponge/util/data/immutable/ImmutableBlockBagData.java
@@ -14,9 +14,10 @@
  * You should have received a copy of the GNU General Public License along with this program. If not,
  * see <http://www.gnu.org/licenses/>.
  */
-package com.sk89q.craftbook.sponge.mechanics.blockbags.data;
+package com.sk89q.craftbook.sponge.util.data.immutable;
 
-import com.sk89q.craftbook.sponge.mechanics.blockbags.BlockBagManager;
+import com.sk89q.craftbook.sponge.util.data.CraftBookKeys;
+import com.sk89q.craftbook.sponge.util.data.mutable.BlockBagData;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.data.DataContainer;
 import org.spongepowered.api.data.manipulator.immutable.common.AbstractImmutableSingleData;
@@ -27,13 +28,13 @@ import org.spongepowered.api.util.annotation.NonnullByDefault;
 public class ImmutableBlockBagData extends AbstractImmutableSingleData<Long, ImmutableBlockBagData, BlockBagData> {
 
     public ImmutableBlockBagData(Long value) {
-        super(value, BlockBagManager.BLOCK_BAG);
+        super(value, CraftBookKeys.BLOCK_BAG);
     }
 
     @Override
     protected ImmutableValue<Long> getValueGetter() {
         return Sponge.getRegistry().getValueFactory()
-                .createValue(BlockBagManager.BLOCK_BAG, getValue())
+                .createValue(CraftBookKeys.BLOCK_BAG, getValue())
                 .asImmutable();
     }
 
@@ -45,7 +46,7 @@ public class ImmutableBlockBagData extends AbstractImmutableSingleData<Long, Imm
     @Override
     public DataContainer toContainer() {
         return super.toContainer()
-                .set(BlockBagManager.BLOCK_BAG, this.getValue());
+                .set(CraftBookKeys.BLOCK_BAG, this.getValue());
     }
 
     @Override
